requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
binance-connector==3.12.0
certifi==2025.6.15
charset-normalizer==3.4.2
dateparser==1.2.2
frozenlist==1.7.0
Flask==3.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.4.2
multidict==6.6.3
numpy==2.3.1
pandas==2.3.0
propcache==0.3.2
pycryptodome==3.23.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.2
pytz==2025.2
PyYAML==6.0.2
psycopg2-binary==2.9.9
regex==2024.11.6
requests==2.32.4
six==1.17.0
scikit-learn==1.8.0
sniffio==1.3.1
ta==0.11.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
